Added mode arg to open() and fixed a scheduling bug

+++ software/src/libc/arch/68k-os/unistd/creat.c
@@ -5,6 +5,6 @@
 
 int creat(const char *path, mode_t mode)
 {
-	return SYSCALL2(SYS_CREAT, path, mode);
+	return SYSCALL2(SYS_CREAT, (int) path, (int) mode);
 }
